+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 _No unreleased changes_
 
+## [1.2.0] - 2021-09-18
+### Added
+- Added handling mentions as commands
+
+### Changed
+- Changed commands to work with discord.js 13
+
 ## [1.1.0] - 2021-09-01
 ### Added
 - Added per guild setting files 
@@ -24,6 +31,7 @@ _No unreleased changes_
 ## [1.0.0] - 2020-04-18
 ### Initial Release
 
+[1.2.0]: https://github.com/VariXx/frogbot/tree/v1.1.0
 [1.1.0]: https://github.com/VariXx/frogbot/tree/v1.1.0
 [1.0.0]: https://github.com/VariXx/frogbot/tree/v1.0.0
 [Unreleased]: https://github.com/VariXx/frogbot/compare/master...develop